[Refactor] Clean up code

+++ src/main/java/com/github/hydos/ginger/engine/font/MetaFile.java
@@ -142,7 +142,7 @@ public class MetaFile
 	{
 		processNextLine();
 		int lineHeightPixels = getValueOfVariable("lineHeight") - paddingHeight;
-		verticalPerPixelSize = TextMeshCreator.LINE_HEIGHT / (double) lineHeightPixels;
+		verticalPerPixelSize = TextMeshCreator.LINE_HEIGHT / lineHeightPixels;
 		horizontalPerPixelSize = verticalPerPixelSize / aspectRatio;
 	}
